Add Tiled boilerplate
