+++ .gitignore
@@ -39,6 +39,7 @@ Ankh.NoLoad
 _ReSharper*/
 *.resharper
 [Tt]est[Rr]esult*
+*.tiled-session
 
 #Project files
 [Bb]uild/
