+++ GateAdventures/Engine/TiledMapLoader.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework.Input;
+
+namespace GateAdventures.Engine
+{
+	public class TiledMapLoader
+	{
+		public TiledMapLoader()
+		{
+			
+		}
+	}
+}
