+++ GateAdventures/Main.cs
@@ -1,0 +1,61 @@
+﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework.Input;
+
+namespace GateAdventures
+{
+	public class Main : Game
+	{
+		private GraphicsDeviceManager _graphics;
+		private SpriteBatch _spriteBatch;
+
+		public Main()
+		{
+			_graphics = new GraphicsDeviceManager(this);
+			Content.RootDirectory = "Content";
+			IsMouseVisible = true;
+		}
+
+		protected override void Initialize()
+		{
+			// TODO: Add your initialization logic here
+
+			base.Initialize();
+		}
+
+		protected override void LoadContent()
+		{
+			_spriteBatch = new SpriteBatch(GraphicsDevice);
+
+			// TODO: use this.Content to load your game content here
+		}
+
+		protected override void Update(GameTime gameTime)
+		{
+			if (GamePad.GetState(PlayerIndex.One).Buttons.Back == ButtonState.Pressed || Keyboard.GetState().IsKeyDown(Keys.Escape))
+				Exit();
+
+			// TODO: Add your update logic here
+
+			base.Update(gameTime);
+		}
+
+		protected override void Draw(GameTime gameTime)
+		{
+			GraphicsDevice.Clear(Color.CornflowerBlue);
+
+			// TODO: Add your drawing code here
+
+			base.Draw(gameTime);
+		}
+	}
+
+	public static class Program
+	{
+		static void Main()
+		{
+			using (var game = new Main())
+				game.Run();
+		}
+	}
+}
