--- GateAdventures/Program.cs
@@ -1,3 +1,0 @@
-﻿
-using var game = new GateAdventures.Game1();
-game.Run();
