+++ Tiled/exampleMap.tmx
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<map version="1.10" tiledversion="1.10.2" orientation="orthogonal" renderorder="right-down" width="30" height="30" tilewidth="32" tileheight="32" infinite="0" nextlayerid="2" nextobjectid="1">
+ <tileset firstgid="1" source="tilesets/pixelArtBasic/TX-Tileset-Grass.tsx"/>
+ <layer id="1" name="Tile Layer 1" width="30" height="30">
+  <data encoding="csv">
+1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,
+1,1,1,1,1,1,30,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,
+1,1,1,1,1,1,1,1,1,1,1,1,1,21,1,1,1,1,1,1,1,1,1,1,1,1,30,1,1,1,
+1,1,1,1,1,1,21,1,1,1,1,1,1,1,1,1,1,1,26,26,1,1,1,1,1,1,1,1,1,1,
+1,1,1,26,26,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,26,30,1,1,1,1,1,1,1,1,
+1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,26,26,1,1,1,1,1,1,1,
+1,1,1,1,1,17,17,17,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,
+17,25,17,17,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,
+17,1,1,1,1,1,26,1,26,26,1,1,1,1,1,1,1,1,1,1,26,26,21,1,1,1,1,1,1,1,
+1,1,1,1,1,1,26,1,1,26,26,1,1,1,1,1,1,1,1,1,1,1,1,1,1,26,1,1,1,1,
+17,1,1,1,1,26,1,1,1,1,1,1,1,1,1,1,1,1,15,1,1,1,1,1,1,1,1,1,1,1,
+17,1,1,15,17,26,21,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,
+17,17,1,1,26,26,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,17,26,26,1,1,
+1,1,1,26,26,1,1,1,1,1,1,1,1,21,21,1,1,1,1,1,1,15,1,1,1,1,1,1,33,33,
+33,33,26,33,33,33,1,1,1,1,1,1,1,1,1,1,26,1,1,1,1,1,1,1,1,1,33,33,33,42,
+33,33,33,33,33,33,33,33,33,33,33,33,33,33,33,20,20,25,26,1,1,1,1,26,1,38,41,33,33,33,
+33,33,33,33,33,33,33,33,33,33,33,33,33,33,33,33,33,1,1,1,1,1,26,26,42,33,42,33,46,26,
+1,1,1,1,1,1,26,46,47,33,33,33,33,33,33,33,33,33,33,1,1,26,26,1,33,33,1,1,26,26,
+1,1,1,1,1,1,1,26,1,1,26,20,1,1,1,1,1,1,33,33,33,33,33,33,33,1,1,1,1,1,
+20,20,1,20,20,20,20,20,20,20,20,20,1,1,1,1,1,1,1,1,33,1,33,33,33,33,1,1,1,1,
+20,20,17,20,20,20,1,1,20,20,20,26,3,3,3,1,1,1,1,1,1,1,1,1,33,33,33,1,1,1,
+17,20,20,15,1,1,1,26,26,3,1,26,26,1,1,1,1,1,1,1,1,1,1,1,33,33,1,1,1,1,
+17,17,17,1,20,20,26,3,3,3,3,3,3,3,3,3,3,3,1,1,1,3,1,33,33,1,1,1,1,1,
+1,1,1,1,1,17,17,1,21,3,3,3,3,3,3,3,3,1,1,1,1,1,1,33,33,33,26,26,26,26,
+1,1,1,17,17,17,15,15,1,25,1,1,1,1,1,1,30,1,1,1,1,1,1,33,42,33,33,1,1,1,
+1,1,1,1,1,1,1,15,15,1,30,1,1,25,1,1,1,1,1,1,26,26,26,1,33,33,33,33,1,1,
+1,17,30,1,1,1,1,26,26,26,26,26,26,26,26,15,15,15,15,15,26,26,26,26,26,33,33,33,1,1,
+17,26,26,17,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,33,42,42,26,
+26,26,1,1,1,1,1,1,1,7,7,1,1,25,17,17,25,1,1,1,1,1,1,1,26,26,26,33,42,33,
+26,26,26,26,26,26,26,26,26,26,26,1,1,1,1,1,1,1,1,1,1,1,1,1,26,26,26,26,33,41
+</data>
+ </layer>
+</map>
