+++ Tiled/tilesets/pixelArtBasic/TX-Tileset-Grass.tsx
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.10.2" name="TX-Tileset-Grass" tilewidth="32" tileheight="32" tilecount="64" columns="8">
+ <image source="TX Tileset Grass.png" width="256" height="256"/>
+</tileset>
